Bring the same concept of "Builder" to collection radio buttons/check boxes
This increases flexibility allowing us to yield only the builder to the
user and hiding specific details of the implementation inside it, also
reusing arguments we already know, such as attribute name and value.

This commit also adds the ability to access the current object/item in
the given collection within the builder instance, giving extra
flexibility to generate custom attributes to label/radio buttons/check
boxes.

Besides, this will make it easy for us to migrate to Rails 4 later as we
are sticking with the same concept here.