Do not lookup for hints if it was explicitly given false
After #405 we added hint classes for the wrappers, but this has forced the
loading of the hint text doing I18n lookups, even though it was explicitly
given false. This checks for the option in #has_hint?, avoiding the lookup
in such cases.

This issues has been caught after #627 and some further investigation,
thanks to @shwoodard.