Fix the way `detect_common_display_methods` finds label & value methods
It was previously looking at an instance of the `collection` instead of the
`collection_classes`. Since the grouped collection input supplies a list
of collection classes that differ from the regular collection, it was
incorrectly looking for labelling and value methods on the option
instances, rather than the optgroup instances.