Refactor test models to always delegate to a fake Relation object
This way both where and order methods are part of the fake relation
object and everything is delegated to that. Also, if we end up changing
the implementation to call some other method the we want this to fail
miserably, and it's better to use a bare object rather than an array to
avoid it passing by mistake.