Small fix on boolean inputs where there are no automatic classes.
I like to keep my markup without a lot of extra hooks that I don't use so
I disable a lot of the automatic classes. This caused a bug when
using a boolean because it assumes there will already be an array of
classes.

This patch removes that assumption by initializing an empty array if
not defined.