Support to date, datetime and time inputs.
I changed `DateTimeInput` to support the new HTML5 form helpers that
comes with Rails 4 (`date_input`, `datetime_input` and `time_input`)
instead of rendering datetime selects:

For example, this input:
```erb
<%= f.input :born_on, as: :date %>
```
Will render:
```html
<input id="user_born_on" name="user[born_on]" type="date" />
```

Datetime selects are still present, but only when HTML5 compatibility
is disabled. It's also possible to render selects even with HTML5 on,
by setting the `html5` option to false.

If we wanted to use an old date select on the example above, we should do:
```erb
<%= f.input :born_on, as: :date, html5: false %>
```

Closes #844.