Allows to use item_wrapper_class with procs
I would like to have this functionality to select a specific check_box with JS.

I would've expected item_wrapper class to behave like label_method and value_method and was surprised it didn't. As each checkbox depends on an item, I expected to be able to derive a class from the item of a check_box as well, instead of assigning the same class to all check_boxes.