Add a way to specify custom collection option lookup key for i18n
A custom key can specified with `config.i18n_options_scope`:

    config.i18n_options_scope = lambda { |i18n_scope| "#{i18n_scope}.options" }

The default (above) returns `simple_form.options` - like it was always
before. But I can no easy override the `options` key like this:

    config.i18n_options_scope = lambda { |i18n_scope| "#{i18n_scope}.other" }

In addition, if I want it to be somewhere else outside the regular
`i18n_scope`, I can do:

    config.i18n_options_scope = lambda { |_| 'where.ever.i.want' }

This gives total flexibility where simple_form looks up translations of
collection options.

The current version always wants the config to be a `lambda`. One could
rewrite the implementation to work with a simple string
(`config.i18n_options_scope = 'where.ever.i.want`) if we want to give
the user total ease of defining the scope in case the scope does not
depend on the `i18n_scope`. Or the implementation could be that in case
of a single string the `i18n_scope` is automatically prepended, for
example:

    config.i18n_options_scope = 'where.ever.i.want'

... means to look up under `simple_form.where.ever.i.want`. But for
explicitness sake requiring it to be a `lambda` always might be the
better choice (but I'm open to changes).