Ensure the arg sent to `type_for_attribute` is a string
Per the docs, it is required to be a string:
http://api.rubyonrails.org/v5.0.0/classes/ActiveRecord/ModelSchema/ClassMethods.html#method-i-type_for_attribute

If you send a symbol, you will get an empty
ActiveRecord::Type::Value instance instead of the actual
column type you might have expected.

Since `f.input [attr]` can otherwise be invoked with
either a symbol or string, it seems important to
support both.

Fixes #1421