Fix error when the scope from association has parameter
As we are using the instance_exec to execute the scope from the
association, when the scope has a parameter, it's breaking.

The scope's parameter always will be the instance of the own class, so
we can fix it using the object instance that simple form uses to build
the form.