Boolean inputs in nested style accepts non-string label
Simple Form isn't rendering the label in boolean nested style when the
input is radio or checkbox with a collection that the first item is
a non-string value.

When Simple Form is building the label tag, it concatenates the input
HTML that is a string with the attribute from a collection that should
be a string too. If it is not, it will convert the attribute to a
character before concatenation and it will not be shown in the view.

**TL;DR**
The input generated by the collection_builder is a `ActiveSupport`
instance. So when the method `+` is called it's calling the
String#concat.

```bash
collection_builder.radio_button + 1
=> "<input class=\"radio_buttons \" id=\"user_amount\" />\u0001"

collection_builder.radio_button + 1.to_s
=> "<input class=\"radio_buttons \" id=\"user_amount\" />1"
```