Remove instruction to use form-inline class
The previous instruction was a bit confusing, as form-inline is only one of the many possible classes that can be used. It’s also (probably) not the one that is most often used, so mentioning it in the installation output implies that it’s more significant than it actually is.

The sentence was also imperative:

	Inside your views, use the 'simple_form_for' with the Bootstrap form class, ‘.form-inline’