Add support for Ruby 3, drop support for Ruby < 2.5, remove I18nCache
I18nCache caused some issues upgrading to Ruby 3, and at the end of the
day I think it's more complexity to keep it around for a marginal gain.

I18n lookups can be expensive, yes, but for most this isn't going to be
a performance hit, and for those where it might be, consider looking at
I18n caching alternatives that aren't specific to these few places where
we had this caching, but to lookups across the board.

Closes #1721
Closes #1720