Run tests with latest gems, specifically country_select v9.0.0
Note that we need to keep Nokogiri on v1.15.6 and Bundler on v2.4.22 due
to Ruby 2.7 compatibility for now.

We also need to lock country_select on v8.x since v9.x will load
countries v6, which does not support Ruby 2.7 anymore either.

(soon we'll have to drop it too apparently!)