Fix error component without object on Rails main to display properly
Rails main changed `object` to be `false` instead of `nil`, causing the
specific check we were doing for `object.nil?` to fail, and the custom
errors were not displaying for forms without an object. Our test suite
caught that, and the fix is to simply not check specifically for `nil?`,
but just whether we have an object or not, which should account for both
`nil`/`false` versions.

It's a small backwards incompatibility, probably not something worth
changing the framework for, but one worth mentioning.

Source is likely to be:
- https://github.com/rails/rails/pull/49943
- https://github.com/rails/rails/pull/50931