Avoid creating multiple regexp objects
The check is slightly different but should work for our purposes here.

Using `end_with?(mapping)` to keep a similar check doesn't work because
some Ruby versions include "did you mean", which make the error message
look something like this:

    NameError: uninitialized constant StringInput

        at.const_get(mapping)
          ^^^^^^^^^^
    Did you mean?  StringInputTest

The regexp check worked because it was matching the end of the line, not
end of string, with `$`, so it'd match the first line of the error
message always.